Postgresql + Flask/requirements.txt
# FastAPI backend
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pydantic>=2.0.0

# Flask frontend
flask>=3.0.0
requests>=2.31.0
